# docs/swagger.yaml
basePath: /
definitions:
  db.Event:
    properties:
      address_line_one:
        type: string
      capacity:
        type: integer
      city:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      event_end:
        type: string
      event_start:
        type: string
      has_rated:
        type: boolean
      host_page_id:
        type: string
      host_user_id:
        type: string
      id:
        type: string
      image_url:
        type: string
      joined:
        type: boolean
      location:
        type: string
      maps_link:
        type: string
      pincode:
        type: string
      price:
        type: integer
      registrant_count:
        type: integer
      tags:
        items:
          $ref: '#/definitions/db.EventTag'
        type: array
      things_provided:
        type: string
      things_to_bring:
        type: string
      title:
        type: string
    type: object
...
  db.EventTag:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  db.HostProfile:
    properties:
      age:
        type: integer
      attendee_rating:
        type: number
      avatar_url:
        type: string
      bio:
        type: string
      city:
        type: string
      email:
        type: string
      gender:
        type: string
      hosting_rating:
        type: number
      name:
        type: string
      past_events:
        items:
          $ref: '#/definitions/db.HostedEvent'
        type: array
      total_attended:
        type: integer
      total_hosted:
        type: integer
      total_ratings:
        type: integer
      user_id:
        type: string
    type: object
  db.HostedEvent:
    properties:
      avg_rating:
        type: number
      city:
        type: string
      event_start:
        type: string
      id:
        type: string
      image_url:
        type: string
      location:
        type: string
      price:
        type: integer
      rating_count:
        type: integer
      title:
        type: string
    type: object
...
host: staging-api.spotlightinfo.in
info:
  contact: {}
  description: REST API for Spotlight — a community events platform
  title: Spotlight API
  version: "1.0"
paths:
  /auth/fcm-token:
    post:
      consumes:
      - application/json
      description: Saves the provided FCM token for the currently authenticated user.
        This token can be used to send push notifications to the user's device.
      parameters:
      - description: Request body containing the FCM token
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: token saved successfully'
          schema:
            type: string
        "400":
          description: 'Bad Request: invalid request body or missing token'
          schema:
            type: string
        "401":
          description: 'Unauthorized: user not authenticated'
          schema:
            type: string
        "500":
          description: 'Internal Server Error: failed to save token'
          schema:
            type: string
      summary: Save a Firebase Cloud Messaging (FCM) token for the authenticated user
      tags:
      - auth
...
  /events/{id}/dismiss-rating-prompt:
    post:
      description: Allows the currently authenticated user to dismiss the rating prompt
        for a specific event. This indicates that the user does not wish to provide
        a rating for that event at this time.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: rating prompt dismissed successfully'
          schema:
            type: string
        "400":
          description: 'Bad Request: invalid event ID'
          schema:
            type: string
        "401":
          description: 'Unauthorized: user not authenticated'
          schema:
            type: string
        "500":
          description: 'Internal Server Error: failed to dismiss rating prompt'
          schema:
            type: string
      summary: Dismiss the rating prompt for an event
      tags:
      - ratings
...
  /events/{id}/rate:
    post:
      consumes:
      - application/json
      description: Allows the currently authenticated user to submit a rating for
        another user (either the host or an attendee) of a specific event. The rating
        includes a score, optional comment, and optional tags.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body containing ratee_id, rating_type, score, comment,
          and tags
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: rating submitted successfully'
          schema:
            type: string
        "400":
          description: 'Bad Request: invalid input or constraints not met'
          schema:
            type: string
        "401":
          description: 'Unauthorized: user not authenticated'
          schema:
            type: string
        "403":
          description: 'Forbidden: user not allowed to rate the specified user'
          schema:
            type: string
        "404":
          description: 'Not Found: event not found'
          schema:
            type: string
        "500":
          description: 'Internal Server Error: failed to submit rating'
          schema:
            type: string
      summary: Submit a rating for an event
      tags:
      - ratings
...
  /events/unrated:
    get:
      description: Returns a list of events that the currently authenticated user
        has attended but not yet rated. This allows users to provide feedback on events
        they have participated in.
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JSON object containing a list of unrated events
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/db.Event'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: 'Internal Server Error: failed to fetch unrated events'
          schema:
            type: string
      summary: Get unrated events for the authenticated user
      tags:
      - ratings
...
  /users/{id}/profile:
    get:
      description: Returns the profile information of a specific host, including their
        ID, name, bio, and avatar URL.
      parameters:
      - description: Host User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JSON object containing the host's profile information
          schema:
            $ref: '#/definitions/db.HostProfile'
        "400":
          description: Invalid user ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: 'Internal Server Error: failed to load profile'
          schema:
            type: string
      summary: Get a host's profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"